pyproject.toml
[tool.poetry]
name = "pychat"
version = "1.0.3"
description = "An easy-to-use Python chat!"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
replit = "^2.0.0"
ansicolors = "^1.1.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
